Branching and Learning
After setting up the background I figured I was done with the outside
world. But when trying to stylize it, I run into the issue of graphics
in Java. Now I dont want to run from Java, but I found it simpler to go
to gimp image manipulation program and edit the background to work with
the widgets I will need to use to set up the lock screen. These changes
were rough efforts to stylize, interrupted by some GIMP work and now
back to it.