Wake and Code
Got up, mom came over, drank some water. Then I proceeded to set up the
ActionListener for the answer combo box and button, linking that to the
RoseLock and linking the MainFrame to the RoseLock so that an unlock
call can transfer up the line and allow the mainframe to do a cardswitch
to a test JPanel. Need to replace the test JPanel with my desired
JPanel.