Orienting myself
I arrive on OpenGL and find that I have no idea of its axis system. I
set up axis the length of the screen, blue ix x, red is y, and z is
green, a circle made before is also green and will remain placed for
convenience to remember to think in three-d. This is sort of what I
think to be my starting point when constructing this damn thing.