Squares is a go
I can now draw squares. I commented the function for future reference.
It takes in the squares side length, the brush for the GLPanel, top left
coordinates and enums for orienting the drawing of the square. Currently
can only draw on planes, does not have angle or rotation component and
can not have more than two dimensions..