I was RIGHT Huzzah!
Ok, so I realized that all the switch statemements were doing was
creating an initial square, which was silly when the loop I put to
create square spanning the proper angle of the given plane created the
initial square as part of its normal operation. I worked long on the
switches, fucking squares man....but glad I realized this and could
eliminate that messy code.